fix(pkg): Rename repo package to client/core
