+++ pkg/client/core/client_test.go
@@ -1,4 +1,4 @@
-package repo
+package core
 
 import (
 	"fmt"
