+++ pkg/client/core/common.go
@@ -1,4 +1,4 @@
-package repo
+package core
 
 import (
 	"bytes"
