+++ pkg/client/core/helmclassic_test.go
@@ -1,4 +1,4 @@
-package repo
+package core
 
 import (
 	"io/ioutil"
@@ -44,7 +44,7 @@ func TestDownloadFromHelmClassic(t *testing.T) {
 	if err != nil {
 		t.Fatalf("error loading index.yaml: %v", err)
 	}
-	if err := sc.DownloadChart(chartPath, "nginx", "5.3.1", sourceHelm.Repo, sourceIndex); err != nil {
+	if err := sc.Fetch(chartPath, "nginx", "5.3.1", sourceHelm.Repo, sourceIndex); err != nil {
 		t.Fatal(err)
 	}
 	if _, err := os.Stat(chartPath); err != nil {
@@ -76,7 +76,7 @@ func TestPublishToHelmClassic(t *testing.T) {
 		t.Fatal("could not create a client for the target repo", err)
 	}
 	chartPath := "../../testdata/apache-7.3.15.tgz"
-	err = tc.PublishChart(chartPath, targetHelm.Repo)
+	err = tc.Push(chartPath, targetHelm.Repo)
 	expectedErrorMsg := "publishing to a Helm classic repository is not supported yet"
 	if err.Error() != expectedErrorMsg {
 		t.Errorf("incorrect error, got: \n %s \n, want: \n %s \n", err.Error(), expectedErrorMsg)
